Add test file b/w 0 and 1 mb to test S3 server.
File is 5 kb.
Currently, on dev/work-rona, put of flood-grid.svg into a created
bucket results in the error "Bucket does not exist" (or something
similar). Upload of 2 MB file does not result in this error. Try-
ing get of flood-grid.svg after failed put results in 404 without
a retry.